src/kio_server/requirements.txt
arrow==1.2.3
Flask==2.2.2
gunicorn==20.1.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
PyMySQL==1.0.2
paho-mqtt==1.6.1
requests==2.28.1
rich==10.11.0
sqlescapy==1.0.1

ipdb==0.13.11